// lae_desktop/src/app/models/person-ref-model.ts
export class PersonRefModel{
	
    personRefId: number;
    personRefFirstname: string;
    personRefLastname: string;
    personRefSex: string;
    personRefTels: string;
    personRefEmail: string;
    personRefDesc: string;
    personRefAddress: string;
    createdBy: string;
    dateCreated: string;
    modifiedBy: string;
    dateModified: string;
}